Some refactor into isolated modules
