+++ api/opportunity/controller.js
@@ -6,6 +6,7 @@ const service = require('./service');
 const notification = require('../notification/service');
 const badgeService = require('../badge/service')(notification);
 const Badge = require('../model/Badge');
+const elasticService = require('../../elastic/service');
 
 var router = new Router();
 
@@ -24,13 +25,13 @@ router.get('/api/task/export', auth.isAdmin, async (ctx, next) => {
 });
 
 router.get('/api/task/reindex', auth.isAdmin, async (ctx, next) => {
-  var tasks = await service.reindexOpportunities();
+  var tasks = await elasticService.reindexOpportunities();
   ctx.body = tasks.length;
 });
 
 router.get('/api/task/search', async (ctx, next) => {
-  var request = service.convertQueryStringToOpportunitiesSearchRequest(ctx.query);
-  var results = await service.searchOpportunities(request);
+  var request = elasticService.convertQueryStringToOpportunitiesSearchRequest(ctx.query);
+  var results = await elasticService.searchOpportunities(request);
 
   ctx.body = { results: results };
 });
