+++ api/opportunity/dao.js
@@ -64,188 +64,6 @@ const taskExportQuery = 'select task.id, task.title, description, task."createdA
   ') as agency_name, task."completedAt" ' +
   'from task inner join midas_user on task."userId" = midas_user.id ';
 
-const tasksToIndex = `select row_to_json(row) as task
-from (
-  select
-    t.id,
-    t.title,
-    t.description,
-    case when t.state = 'in progress' and t.accepting_applicants then 'open' else t.state end as "state",
-    t.details,
-    t.outcome,
-    t.about,
-    t.restrict ->> 'abbr' as "restrictedToAgency",
-    u.name,
-    t.accepting_applicants as "acceptingApplicants",
-    (
-      select
-        tags.data ->> 'abbr' as "postingAgency"
-      from tagentity_users__user_tags utags
-        inner join tagentity tags on utags.tagentity_users = tags.id
-      where
-        tags.type = 'agency' and
-        utags.user_tags = u.id
-      order by "updatedAt" desc
-      limit 1
-    ),
-    (
-      select  
-        array_to_json(array_agg(row_to_json(s)))
-      from (
-        select
-          skilltags.id,
-          skilltags.name
-        from tagentity_tasks__task_tags tt
-          left join tagentity skilltags on skilltags.id = tt.tagentity_tasks
-        where
-          skilltags.type = 'task-people' and
-          tt."task_tags" = t.id
-      ) s
-    ) as "task-people",
-    (
-      select
-        array_to_json(array_agg(row_to_json(s)))
-      from (
-        select
-          taskpeopletags.id,
-          taskpeopletags.name
-        from tagentity_tasks__task_tags tt
-          left join tagentity taskpeopletags on taskpeopletags.id = tt.tagentity_tasks
-        where
-          taskpeopletags.type = 'skill' and
-          tt."task_tags" = t.id
-      ) s
-    ) as skills,
-    (
-      select
-        array_to_json(array_agg(row_to_json(s)))
-      from (
-        select
-          tasklengthtags.id,
-          tasklengthtags.name
-        from tagentity_tasks__task_tags tt
-          left join tagentity tasklengthtags on tasklengthtags.id = tt.tagentity_tasks
-        where
-          tasklengthtags.type = 'task-length' and
-          tt."task_tags" = t.id
-      ) s
-    ) as "task-length",
-    (
-      select
-        array_to_json(array_agg(row_to_json(l)))
-      from (
-        select
-          languagetags.id,
-          languagetags.name
-        from tagentity_tasks__task_tags tt
-          left join tagentity languagetags on languagetags.id = tt.tagentity_tasks
-        where
-          languagetags.type = 'location' and
-          tt."task_tags" = t.id
-      ) l
-    ) as "location",
-    (
-      select
-        array_to_json(array_agg(row_to_json(l)))
-      from (
-        select
-          locationtags.id,
-          locationtags.name
-        from tagentity_tasks__task_tags tt
-          left join tagentity locationtags on locationtags.id = tt.tagentity_tasks
-        where
-          locationtags.type = 'task-time-estimate' and
-          tt."task_tags" = t.id
-      ) l
-    ) as "task-time-estimate",
-    (
-      select
-        array_to_json(array_agg(row_to_json(l)))
-      from (
-        select
-          ltags.id,
-          ltags.name
-        from tagentity_tasks__task_tags tt
-          left join tagentity ltags on ltags.id = tt.tagentity_tasks
-        where
-          ltags.type = 'series' and
-           tt."task_tags" = t.id
-      ) l
-    ) as series,
-    (
-      select
-        array_to_json(array_agg(row_to_json(l)))
-      from (
-        select
-          timetags.id,
-          timetags.name
-        from tagentity_tasks__task_tags tt
-          left join tagentity timetags on timetags.id = tt.tagentity_tasks
-        where
-          timetags.type = 'task-time-required' and
-          tt."task_tags" = t.id
-      ) l
-    ) as "task-time-required",
-    (
-      select
-        array_to_json(array_agg(row_to_json(l)))
-      from (
-        select
-          careertags.id,
-          careertags.name
-        from tagentity_tasks__task_tags tt
-          left join tagentity careertags on careertags.id = tt.tagentity_tasks
-        where
-          careertags.type = 'career' and
-          tt."task_tags" = t.id
-      ) l
-    ) as "career",
-    (
-      select array_to_json(array_agg(row_to_json(s)))
-      from (
-        select
-          skilltags.id,
-          skilltags.name
-        from tagentity_tasks__task_tags tt
-          left join tagentity skilltags on skilltags.id = tt.tagentity_tasks
-        where
-          skilltags.type = 'task-skills-required'
-          and tt."task_tags" = t.id
-      ) s
-    ) as "task-skills-required",
-    (
-      select 
-        array_to_json(array_agg(row_to_json(k)))
-      from (
-        select 
-          keytags.id,
-          keytags.name
-        from tagentity_tasks__task_tags tt
-          left join tagentity keytags on keytags.id = tt.tagentity_tasks
-        where
-         keytags.type = 'keywords' and
-         tt."task_tags" = t.id
-      ) k
-    ) as keywords,
-    (
-      select
-       array_to_json(array_agg(row_to_json(k)))
-      from (
-        select
-          topictags.id,
-          topictags.name
-        from tagentity_tasks__task_tags tt
-          left join tagentity topictags on topictags.id = tt.tagentity_tasks
-        where
-          topictags.type = 'topic' and
-          tt."task_tags" = t.id
-      ) k
-    ) as topic
-  from task t
-    inner join midas_user u on t."userId" = u.id         
-  %s
-) row`;
-
 var exportFormat = {
   'task_id': 'id',
   'name': {field: 'title', filter: nullToEmptyString},
@@ -357,8 +175,6 @@ module.exports = function (db) {
       userTasks: userTasksQuery,
       tasksDueQuery: tasksDueQuery,
       tasksDueDetailQuery: tasksDueDetailQuery,
-      tasksToIndex: util.format(tasksToIndex,"order by t.id"),
-      taskToIndex: util.format(tasksToIndex,"where t.id = ?")
     },
     options: options,
     clean: clean,
