+++ api/opportunity/service.js
@@ -1,7 +1,7 @@
 const _ = require ('lodash');
 const log = require('log')('app:opportunity:service');
 const db = require('../../db');
-const elasticClient = require('../../elastic');
+const elasticService = require('../../elastic/service');
 const dao = require('./dao')(db);
 const notification = require('../notification/service');
 const badgeService = require('../badge/service')(notification);
@@ -111,7 +111,7 @@ async function createOpportunity (attributes, done) {
       task.tags = tags;
     });
     task.owner = dao.clean.user((await dao.User.query(dao.query.user, task.userId, dao.options.user))[0]);
-    await indexOpportunity(task.id);
+    await elasticService.indexOpportunity(task.id);
     return done(null, task);
   }).catch(err => {
     return done(true);
@@ -165,7 +165,7 @@ async function updateOpportunityState (attributes, done) {
   await dao.Task.update(attributes).then(async (t) => {
     var task = await findById(t.id, true);
     task.previousState = origTask.state;
-    await indexOpportunity(task.id);
+    await elasticService.indexOpportunity(task.id);
     return done(task, origTask.state !== task.state);
   }).catch (err => {
     return done(null, false, {'message':'Error updating task.'});
@@ -194,7 +194,7 @@ async function updateOpportunity (attributes, done) {
       .then(async () => {
         await processTaskTags(task, tags).then(async tags => {
           task.tags = tags;
-          await indexOpportunity(task.id);
+          await elasticService.indexOpportunity(task.id);
           return done(task, origTask.state !== task.state);
         });
       }).catch (err => { return done(null, false, {'message':'Error updating task.'}); });
@@ -209,7 +209,7 @@ async function publishTask (attributes, done) {
   await dao.Task.update(attributes).then(async (t) => {
     var task = await findById(t.id, true);
     sendTaskNotification(task.owner, task, 'task.update.opened');
-    await indexOpportunity(task.id);
+    await elasticService.indexOpportunity(task.id);
     return done(true);
   }).catch (err => {
     return done(false);
@@ -350,7 +350,7 @@ async function copyOpportunity (attributes, user, done) {
           log.info('register: failed to update tag ', attributes.username, tag, err);
         });
       });
-      await indexOpportunity(task.id);
+      await elasticService.indexOpportunity(task.id);
       return done(null, { 'taskId': task.id });
     }).catch (err => { return done({'message':'Error copying task.'}); });
 }
@@ -372,7 +372,7 @@ async function deleteTask (id) {
   await dao.TaskTags.delete('task_tags = ?', id).then(async (task) => {
     dao.Volunteer.delete('"taskId" = ?', id).then(async (task) => {
       dao.Task.delete('id = ?', id).then(async (task) => {
-        await indexOpportunity(id);
+        await elasticService.indexOpportunity(id);
         return id;
       }).catch(err => {
         log.info('delete: failed to delete task ', err);
@@ -407,131 +407,6 @@ async function getExportData () {
     fields: fields,
     fieldNames: fieldNames,
   });
-}
-
-async function reindexOpportunities () {
-  var records =  _.map((await dao.Task.db.query(dao.query.tasksToIndex)).rows, toElasticOpportunity);
-
-  var bulk_request = [];
-
-  for(i=0; i<records.length; i++){
-    bulk_request.push({index: { _index: 'task', _type: 'task', _id: records[i].id }});
-    bulk_request.push(records[i]);
-  }
-
-  await elasticClient.bulk({ body: bulk_request });
-  return records;
-}
-
-async function searchOpportunities (request) {
-  var results = null;
-  if(request){
-    results = await elasticClient.search({
-      index: 'task',
-      type: 'task',
-      body: JSON.stringify(request),
-    });
-  }
-  else{
-    results = await elasticClient.search({ index: 'task' });
-  }
-  
-  return results;
-}
-
-function convertQueryStringToOpportunitiesSearchRequest (query){
-  var request = {
-    query : {
-      bool : {
-        filter : {
-          bool: {
-            must: [] ,
-            must_not: [],
-          },
-        },
-      },
-    },
-    addTerms: function (filter, field) { 
-      if(filter){
-        filter_must.push({terms: { [field] : asArray(filter) }});
-      }
-    }
-  };
-  var filter_must = request.query.bool.filter.bool.must;
-  var filter_must_not = request.query.bool.filter.bool.must_not;
-  request.addTerms(query.status, 'state' );
-  request.addTerms(query.skill, 'skills.name' );
-  request.addTerms(query.career, 'careers.name' );
-  request.addTerms(query.series, 'series.code' );
-  request.addTerms(query.timerequired, 'timeRequired' );
-  request.addTerms(query.locationtype, 'locationType' );
-  request.addTerms(query.restrictedtoagency, 'restrictedToAgency' );
-  if(!('includerestricted' in query)){
-    filter_must_not.push({exists: { field: 'restrictedToAgency' }});
-  }
-  if(query.keyword){
-    var keyword = '';
-    keyword = Array.isArray(query.keyword) ?query.keyword.join(' ') : query.keyword;
-    request.query.bool.must = {
-      simple_query_string: {
-        query : keyword,
-      }
-    };
-
-  }
-  
-  //f(query.keyword) request.keyword = query.keyword;
-  console.log(request);
-  return request;
-}
-
-async function indexOpportunity (taskId) {
-  if (!(await elasticClient.IsAlive()))
-  {
-    return null;
-  }
-  var records = _.map((await dao.Task.db.query(dao.query.taskToIndex, taskId)).rows, toElasticOpportunity);
-
-  var bulk_request = [];
-
-  for(i=0; i<records.length; i++){
-    bulk_request.push({index: { _index: 'task', _type: 'task', _id: records[i].id }});
-    bulk_request.push(records[i]);
-  }
-
-  await elasticClient.bulk({ body: bulk_request });
-  return records;
-}
-
-function toElasticOpportunity (value, index, list) {
-  var doc = value.task;
-  return {
-    'id': doc.id,
-    'title': doc.title,
-    'description': doc.description,
-    'state': doc.state,
-    'details': doc.details,
-    'outcome': doc.outcome,
-    'about': doc.about,
-    'restrictedToAgency': doc.restrictedToAgency,
-    'requestor': doc.name,
-    'postingAgency': doc.postingAgency,
-    'acceptingApplicants': doc.acceptingApplicants,
-    'taskPeople': (_.first(doc['task-people']) || { name: null }).name,
-    'timeRequired': (_.first(doc['task-time-required']) || { name: null }).name,
-    'timeEstimate': (_.first(doc['task-time-estimate']) || { name: null }).name,
-    'taskLength': (_.first(doc['task-length']) || { name: null }).name, 
-    'skills': doc.skills,
-    'locationType': Array.isArray(doc.location) && doc.location.length > 0 ? 'In Person' : 'Virtual',
-    'locations': doc.location,
-    'series': _.map(doc.series, (item) => { return  {id: item.id || 0, code: item.name.substring(0,4), name: item.name.replace(/.*\((.*)\).*/,'$1') };}),
-    'careers': doc.career,
-    'keywords': _.map(doc.keywords, (item) => item.name),
-  };
-}
-
-function asArray (value) {
-  return Array.isArray(value) ? value: [value];
 }
 
 async function sendTasksDueNotifications (action, i) {
@@ -575,7 +450,6 @@ module.exports = {
   copyOpportunity: copyOpportunity,
   deleteTask: deleteTask,
   getExportData: getExportData,
-  reindexOpportunities: reindexOpportunities,
   volunteersCompleted: volunteersCompleted,
   sendTaskNotification: sendTaskNotification,
   sendTaskStateUpdateNotification: sendTaskStateUpdateNotification,
@@ -584,7 +458,4 @@ module.exports = {
   sendTasksDueNotifications: sendTasksDueNotifications,
   canUpdateOpportunity: canUpdateOpportunity,
   canAdministerTask: canAdministerTask,
-  indexOpportunity: indexOpportunity,
-  convertQueryStringToOpportunitiesSearchRequest: convertQueryStringToOpportunitiesSearchRequest,
-  searchOpportunities: searchOpportunities,
 };
