+++ app/openopps-setup.js
@@ -12,6 +12,9 @@ if (process.env.NEW_RELIC_APP_NAME && process.env.NEW_RELIC_LICENSE_KEY) {
   require('newrelic');
 }
 
+var psqlConnection = appEnv.getServiceCreds('psql-openopps');
+if(psqlConnection) process.env.DB_CONNECTION = psqlConnection;
+
 // get elastic url if it exists from Cloud Foundry service
 var url = appEnv.getServiceURL('elastic56-openopps');
 if (url) process.env.ELASTIC_URL = url;
@@ -27,5 +30,6 @@ _.extend(configuration, require('../config/version'));
 _.extend(configuration, require('../config/fileStore'));
 _.extend(configuration, require('../config/email'));
 _.extend(configuration, require('../elastic/configure-elastic'));
+_.extend(configuration, require('../db/configure-db'));
 
 module.exports = configuration;
